Extract jitting tests into separate test class
